movieRater/requirements.txt
Django==2.2.1
django-crontab==0.7.1
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
